feat:  standalone module
- we create a new module schema_registry

- we move all FastAPI related code into this new folder

- going forward, move out all SR related things into this new folder
