+++ container/karapace.env
@@ -1,0 +1,46 @@
+KARAPACE_DOTENV=/opt/karapace/karapace.env
+ACCESS_LOGS_DEBUG=False
+ADVERTISED_HOSTNAME=karapace-schema-registry
+ADVERTISED_PORT=8081
+ADVERTISED_PROTOCOL=http
+BOOTSTRAP_URI=kafka:29092
+CLIENT_ID=karapace-schema-registry
+COMPATIBILITY=BACKWARD
+CONNECTIONS_MAX_IDLE_MS=15000
+CONSUMER_ENABLE_AUTO_COMMIT=True
+CONSUMER_REQUEST_TIMEOUT_MS=11000
+CONSUMER_REQUEST_MAX_BYTES=67108864
+CONSUMER_IDLE_DISCONNECT_TIMEOUT=0
+FETCH_MIN_BYTES=1
+GROUP_ID=karapace-schema-registry
+HOST=0.0.0.0
+PORT=8081
+REGISTRY_HOST=karapace-registry
+REGISTRY_PORT=8081
+REST_AUTHORIZATION=False
+LOG_HANDLER=stdout
+LOG_LEVEL=DEBUG
+LOG_FORMAT=%(asctime)s [%(threadName)s] %(filename)s:%(funcName)s:%(lineno)d %(message)s
+MASTER_ELIGIBILITY=True
+REPLICATION_FACTOR=1
+SECURITY_PROTOCOL=PLAINTEXT
+SSL_CHECK_HOSTNAME=True
+TOPIC_NAME=_schemas
+METADATA_MAX_AGE_MS=60000
+ADMIN_METADATA_MAX_AGE=5
+PRODUCER_ACKS=1
+PRODUCER_COUNT=5
+PRODUCER_LINGER_MS=100
+PRODUCER_MAX_REQUEST_SIZE=1048576
+SESSION_TIMEOUT_MS=10000
+KARAPACE_REST=False
+KARAPACE_REGISTRY=True
+NAME_STRATEGY=topic_name
+NAME_STRATEGY_VALIDATION=True
+MASTER_ELECTION_STRATEGY=lowest
+PROTOBUF_RUNTIME_DIRECTORY=runtime
+STATSD_HOST=statsd-exporter
+STATSD_PORT=8125
+KAFKA_SCHEMA_READER_STRICT_MODE=False
+KAFKA_RETRIABLE_ERRORS_SILENCED=True
+USE_PROTOBUF_FORMATTER=False
