+++ requirements/requirements-typing.txt
@@ -8,15 +8,20 @@ accept-types==0.4.1
     # via karapace (/karapace/pyproject.toml)
 aiohappyeyeballs==2.4.3
     # via aiohttp
-aiohttp==3.10.10
+aiohttp==3.10.11
     # via karapace (/karapace/pyproject.toml)
 aiokafka==0.10.0
     # via karapace (/karapace/pyproject.toml)
 aiosignal==1.3.1
     # via aiohttp
 anyio==4.6.2.post1
-    # via watchfiles
-async-timeout==4.0.3
+    # via
+    #   httpx
+    #   starlette
+    #   watchfiles
+asgiref==3.8.1
+    # via opentelemetry-instrumentation-asgi
+async-timeout==5.0.1
     # via
     #   aiohttp
     #   aiokafka
@@ -30,23 +35,60 @@ avro @ https://github.com/aiven/avro/archive/5a82d57f2a650fd87c819a30e433f1abb2c
 cachetools==5.3.3
     # via karapace (/karapace/pyproject.toml)
 certifi==2024.8.30
-    # via sentry-sdk
+    # via
+    #   httpcore
+    #   httpx
+    #   sentry-sdk
+click==8.1.7
+    # via
+    #   typer
+    #   uvicorn
 confluent-kafka==2.4.0
     # via karapace (/karapace/pyproject.toml)
 cramjam==2.9.0
     # via python-snappy
+dependency-injector==4.43.0
+    # via karapace (/karapace/pyproject.toml)
+deprecated==1.2.15
+    # via
+    #   opentelemetry-api
+    #   opentelemetry-semantic-conventions
+dnspython==2.7.0
+    # via email-validator
+email-validator==2.2.0
+    # via fastapi
 exceptiongroup==1.2.2
     # via anyio
+fastapi[standard]==0.115.5
+    # via karapace (/karapace/pyproject.toml)
+fastapi-cli[standard]==0.0.5
+    # via fastapi
 frozenlist==1.5.0
     # via
     #   aiohttp
     #   aiosignal
+h11==0.14.0
+    # via
+    #   httpcore
+    #   uvicorn
+httpcore==1.0.7
+    # via httpx
+httptools==0.6.4
+    # via uvicorn
+httpx==0.27.2
+    # via fastapi
 idna==3.10
     # via
     #   anyio
+    #   email-validator
+    #   httpx
     #   yarl
+importlib-metadata==8.5.0
+    # via opentelemetry-api
 isodate==0.7.2
     # via karapace (/karapace/pyproject.toml)
+jinja2==3.1.4
+    # via fastapi
 jsonschema==4.23.0
     # via karapace (/karapace/pyproject.toml)
 jsonschema-specifications==2024.10.1
@@ -55,6 +97,8 @@ lz4==4.3.3
     # via karapace (/karapace/pyproject.toml)
 markdown-it-py==3.0.0
     # via rich
+markupsafe==3.0.2
+    # via jinja2
 mdurl==0.1.2
     # via markdown-it-py
 multidict==6.1.0
@@ -67,41 +111,93 @@ mypy-extensions==1.0.0
     # via mypy
 networkx==3.2.1
     # via karapace (/karapace/pyproject.toml)
+opentelemetry-api==1.28.2
+    # via
+    #   karapace (/karapace/pyproject.toml)
+    #   opentelemetry-instrumentation
+    #   opentelemetry-instrumentation-asgi
+    #   opentelemetry-instrumentation-fastapi
+    #   opentelemetry-sdk
+    #   opentelemetry-semantic-conventions
+opentelemetry-instrumentation==0.49b2
+    # via
+    #   opentelemetry-instrumentation-asgi
+    #   opentelemetry-instrumentation-fastapi
+opentelemetry-instrumentation-asgi==0.49b2
+    # via opentelemetry-instrumentation-fastapi
+opentelemetry-instrumentation-fastapi==0.49b2
+    # via karapace (/karapace/pyproject.toml)
+opentelemetry-sdk==1.28.2
+    # via karapace (/karapace/pyproject.toml)
+opentelemetry-semantic-conventions==0.49b2
+    # via
+    #   opentelemetry-instrumentation
+    #   opentelemetry-instrumentation-asgi
+    #   opentelemetry-instrumentation-fastapi
+    #   opentelemetry-sdk
+opentelemetry-util-http==0.49b2
+    # via
+    #   opentelemetry-instrumentation-asgi
+    #   opentelemetry-instrumentation-fastapi
 packaging==24.2
-    # via aiokafka
+    # via
+    #   aiokafka
+    #   opentelemetry-instrumentation
 prometheus-client==0.20.0
     # via karapace (/karapace/pyproject.toml)
 protobuf==3.20.3
     # via karapace (/karapace/pyproject.toml)
+pydantic==1.10.17
+    # via
+    #   fastapi
+    #   karapace (/karapace/pyproject.toml)
 pygments==2.18.0
     # via rich
-pyjwt==2.9.0
+pyjwt==2.10.0
     # via karapace (/karapace/pyproject.toml)
 python-dateutil==2.9.0.post0
     # via karapace (/karapace/pyproject.toml)
+python-dotenv==1.0.1
+    # via uvicorn
+python-multipart==0.0.17
+    # via fastapi
 python-snappy==0.7.3
     # via karapace (/karapace/pyproject.toml)
+pyyaml==6.0.2
+    # via uvicorn
 referencing==0.35.1
     # via
     #   jsonschema
     #   jsonschema-specifications
     #   types-jsonschema
 rich==13.7.1
-    # via karapace (/karapace/pyproject.toml)
+    # via
+    #   karapace (/karapace/pyproject.toml)
+    #   typer
 rpds-py==0.21.0
     # via
     #   jsonschema
     #   referencing
-sentry-sdk==2.18.0
+sentry-sdk==2.19.0
     # via karapace (/karapace/pyproject.toml)
+shellingham==1.5.4
+    # via typer
 six==1.16.0
-    # via python-dateutil
+    # via
+    #   dependency-injector
+    #   python-dateutil
 sniffio==1.3.1
-    # via anyio
+    # via
+    #   anyio
+    #   httpx
+starlette==0.41.3
+    # via fastapi
 tenacity==9.0.0
     # via karapace (/karapace/pyproject.toml)
 tomli==2.1.0
     # via mypy
+typer==0.13.1
+    # via fastapi-cli
 types-cachetools==5.5.0.20240820
     # via karapace (/karapace/pyproject.toml)
 types-jsonschema==4.23.0.20240813
@@ -111,20 +207,43 @@ types-protobuf==3.20.4.6
 typing-extensions==4.12.2
     # via
     #   anyio
+    #   asgiref
+    #   fastapi
     #   karapace (/karapace/pyproject.toml)
     #   multidict
     #   mypy
+    #   opentelemetry-sdk
+    #   pydantic
+    #   starlette
+    #   typer
+    #   uvicorn
 ujson==5.10.0
     # via karapace (/karapace/pyproject.toml)
 urllib3==2.2.3
     # via sentry-sdk
+uvicorn[standard]==0.32.1
+    # via
+    #   fastapi
+    #   fastapi-cli
+uvloop==0.21.0
+    # via uvicorn
 watchfiles==0.24.0
-    # via karapace (/karapace/pyproject.toml)
+    # via
+    #   karapace (/karapace/pyproject.toml)
+    #   uvicorn
+websockets==14.1
+    # via uvicorn
+wrapt==1.17.0
+    # via
+    #   deprecated
+    #   opentelemetry-instrumentation
 xxhash==3.5.0
     # via karapace (/karapace/pyproject.toml)
 yarl==1.12.1
     # via
     #   aiohttp
     #   karapace (/karapace/pyproject.toml)
+zipp==3.21.0
+    # via importlib-metadata
 zstandard==0.23.0
     # via karapace (/karapace/pyproject.toml)
