+++ src/karapace/dependencies/controller_dependency.py
@@ -8,7 +8,7 @@ from fastapi import Depends
 from karapace.dependencies.config_dependency import ConfigDep
 from karapace.dependencies.schema_registry_dependency import SchemaRegistryDep
 from karapace.dependencies.stats_dependeny import StatsDep
-from karapace.schema_registry_apis import KarapaceSchemaRegistryController
+from schema_registry.schema_registry_apis import KarapaceSchemaRegistryController
 from typing import Annotated
