+++ src/schema_registry/container.py
@@ -1,0 +1,4 @@
+"""
+Copyright (c) 2024 Aiven Ltd
+See LICENSE for details
+"""
