+++ src/schema_registry/routers/health_router.py
