+++ src/schema_registry/routers/mode_router.py
@@ -7,27 +7,36 @@ from fastapi import APIRouter
 from karapace.auth.auth import Operation
 from karapace.auth.dependencies import AuthenticatorAndAuthorizerDep, CurrentUserDep
 from karapace.dependencies.controller_dependency import KarapaceSchemaRegistryControllerDep
-from karapace.routers.errors import unauthorized
-from karapace.routers.requests import CompatibilityCheckResponse, SchemaRequest
 from karapace.typing import Subject
+from schema_registry.routers.errors import unauthorized
 
-compatibility_router = APIRouter(
-    prefix="/compatibility",
-    tags=["compatibility"],
+mode_router = APIRouter(
+    prefix="/mode",
+    tags=["mode"],
     responses={404: {"description": "Not found"}},
 )
 
 
-@compatibility_router.post("/subjects/{subject}/versions/{version}", response_model_exclude_none=True)
-async def compatibility_post(
+@mode_router.get("")
+async def mode_get(
+    controller: KarapaceSchemaRegistryControllerDep,
+    user: CurrentUserDep,
+    authorizer: AuthenticatorAndAuthorizerDep,
+):
+    if authorizer and not authorizer.check_authorization(user, Operation.Read, "Config:"):
+        raise unauthorized()
+
+    return await controller.get_global_mode()
+
+
+@mode_router.get("/{subject}")
+async def mode_get_subject(
     controller: KarapaceSchemaRegistryControllerDep,
     user: CurrentUserDep,
     authorizer: AuthenticatorAndAuthorizerDep,
     subject: Subject,
-    version: str,  # TODO support actual Version object
-    schema_request: SchemaRequest,
-) -> CompatibilityCheckResponse:
+):
     if authorizer and not authorizer.check_authorization(user, Operation.Read, f"Subject:{subject}"):
         raise unauthorized()
 
-    return await controller.compatibility_check(subject=subject, schema_request=schema_request, version=version)
+    return await controller.get_subject_mode(subject=subject)
