+++ tests/integration/test_schema.py
@@ -9,9 +9,9 @@ from http import HTTPStatus
 from karapace.client import Client
 from karapace.kafka.producer import KafkaProducer
 from karapace.rapu import is_success
-from karapace.schema_registry_apis import SchemaErrorMessages
 from karapace.schema_type import SchemaType
 from karapace.utils import json_encode
+from schema_registry.schema_registry_apis import SchemaErrorMessages
 from tests.base_testcase import BaseTestCase
 from tests.integration.utils.cluster import RegistryDescription
 from tests.integration.utils.kafka_server import KafkaServers
